Reorganized instruction formats into separate module
