+++ name-core/src/instruction/formats/mod.rs
@@ -1,0 +1,10 @@
+pub mod cop_mov_r_type;
+pub mod fp_cc_branch_type;
+pub mod fp_cc_type;
+pub mod fp_four_reg_type;
+pub mod fp_r_type;
+pub mod i_type;
+pub mod j_type;
+pub mod r_type;
+pub mod regimm_i_type;
+mod stub;
