+++ name-core/src/instruction/instruction.rs
@@ -1,9 +1,9 @@
 use std::{fmt, io, path::PathBuf};
 
-use super::information::{ArgumentType, FpInstructionInformation, InstructionInformation};
+use super::information::{FpInstructionInformation, InstructionInformation};
 use crate::{
     instruction::information::FpFmt,
-    parse::{parse::AstKind, span::Span},
+    parse::span::Span,
     structs::ProgramState,
 };
 
@@ -222,627 +222,21 @@ impl RawInstruction {
 }
 
 // RawInstruction to XArgs conversion
-impl From<IArgs> for RawInstruction {
-    fn from(i_args: IArgs) -> Self {
-        RawInstruction::new(
-            (i_args.opcode << 26) | ((i_args.rs) << 21) | ((i_args.rt) << 16) | (i_args.imm as u32),
-        )
-    }
-}
-
-impl From<JArgs> for RawInstruction {
-    fn from(j_args: JArgs) -> Self {
-        RawInstruction::new((j_args.opcode << 26) | (j_args.address))
-    }
-}
-
-impl From<RArgs> for RawInstruction {
-    fn from(r_args: RArgs) -> Self {
-        RawInstruction::new(
-            (r_args.opcode << 26)
-                | ((r_args.rs) << 21)
-                | ((r_args.rt) << 16)
-                | ((r_args.rd) << 11)
-                | ((r_args.shamt) << 6)
-                | (r_args.funct),
-        )
-    }
-}
-
-impl From<FpCCArgs> for RawInstruction {
-    fn from(fp_cc_args: FpCCArgs) -> Self {
-        RawInstruction::new(
-            (fp_cc_args.opcode << 26)
-                | ((fp_cc_args.fmt) << 21)
-                | ((fp_cc_args.ft) << 16)
-                | ((fp_cc_args.fs) << 11)
-                | ((fp_cc_args.cc) << 8)    // Note that this is 8, not 6.
-                | (fp_cc_args.funct),
-        )
-    }
-}
-
-impl From<FpCCBranchArgs> for RawInstruction {
-    fn from(fp_cc_branch_args: FpCCBranchArgs) -> Self {
-        RawInstruction::new(
-            (fp_cc_branch_args.opcode << 26)
-                | ((fp_cc_branch_args.funky_funct) << 21)
-                | ((fp_cc_branch_args.cc) << 18)
-                | ((fp_cc_branch_args.nd) << 17)
-                | ((fp_cc_branch_args.tf) << 16)
-                | (fp_cc_branch_args.offset as u32),
-        )
-    }
-}
-
-impl From<FpRArgs> for RawInstruction {
-    fn from(fp_r_args: FpRArgs) -> Self {
-        RawInstruction::new(
-            (fp_r_args.opcode << 26)
-                | ((fp_r_args.fmt) << 21)
-                | ((fp_r_args.ft) << 16)
-                | ((fp_r_args.fs) << 11)
-                | ((fp_r_args.fd) << 6)
-                | (fp_r_args.funct),
-        )
-    }
-}
-
-impl From<FpFourRegArgs> for RawInstruction {
-    fn from(fp_four_reg_args: FpFourRegArgs) -> Self {
-        RawInstruction::new(
-            (fp_four_reg_args.op_code << 26)
-                | ((fp_four_reg_args.fr << 21))
-                | ((fp_four_reg_args.ft << 16))
-                | ((fp_four_reg_args.fs << 11))
-                | ((fp_four_reg_args.fd << 6))
-                | ((fp_four_reg_args.op4 << 3))
-                | (fp_four_reg_args.fmt3)
-        )
-    }
-}
-
-impl From<CopMovRArgs> for RawInstruction {
-    fn from(cop_mov_r_args: CopMovRArgs) -> Self {
-        RawInstruction::new(
-            (cop_mov_r_args.op_code << 26)
-                | ((cop_mov_r_args.funct_code << 21))
-                | ((cop_mov_r_args.rt << 16))
-                | ((cop_mov_r_args.rd << 11))
-                | (cop_mov_r_args.sel)
-        )
-    }
-}
 
 // IArgs
-#[derive(Debug)]
-pub struct IArgs {
-    pub opcode: u32,
-    pub rs: u32,
-    pub rt: u32,
-    pub imm: u16,
-}
-
-impl IArgs {
-    pub fn assign_i_type_arguments(
-        arguments: Vec<AstKind>,
-        args_to_use: &[ArgumentType],
-    ) -> AssembleResult<Self> {
-        let mut rs: u32 = 0;
-        let mut rt: u32 = 0;
-        let mut imm: u32 = 0;
-
-        for (i, passed) in arguments.into_iter().enumerate() {
-            match args_to_use[i] {
-                ArgumentType::Rs | ArgumentType::Fs => {
-                    rs = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)?
-                }
-                ArgumentType::Rt | ArgumentType::Ft => {
-                    rt = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)?
-                }
-                ArgumentType::Immediate => imm = passed.get_immediate().unwrap_or(0),
-                ArgumentType::Identifier | ArgumentType::BranchLabel => (),
-                _ => unreachable!(),
-            }
-        }
-
-        // Check if the extracted immediate falls within valid range
-        if ((imm as i32) as i16) as u16 != imm as u16 {
-            return Err(ErrorKind::ImmediateOverflow(imm));
-        }
-
-        return Ok(Self {
-            opcode: 0,
-            rs,
-            rt,
-            imm: imm as u16,
-        });
-    }
-}
-
-impl From<RawInstruction> for IArgs {
-    fn from(raw: RawInstruction) -> IArgs {
-        IArgs {
-            opcode: raw.get_opcode(),
-            rs: raw.get_rs(),
-            rt: raw.get_rt(),
-            imm: raw.get_immediate(),
-        }
-    }
-}
 
 // JArgs
-pub struct JArgs {
-    pub opcode: u32,
-    pub address: u32,
-}
-
-impl JArgs {
-    pub fn assign_j_type_arguments(
-        _arguments: Vec<AstKind>,
-        _args_to_use: &[ArgumentType],
-    ) -> AssembleResult<Self> {
-        // Nothing ever happens...
-        Ok(Self {
-            opcode: 0,  // Will be filled in by caller
-            address: 0, // Will be handled by linker
-        })
-    }
-}
-
-impl From<RawInstruction> for JArgs {
-    fn from(raw: RawInstruction) -> JArgs {
-        JArgs {
-            opcode: raw.get_opcode(),
-            address: raw.get_jump(),
-        }
-    }
-}
 
 // RArgs
-#[derive(Debug)]
-pub struct RArgs {
-    pub opcode: u32,
-    pub rs: u32,
-    pub rt: u32,
-    pub rd: u32,
-    pub shamt: u32,
-    pub funct: u32,
-}
-
-impl RArgs {
-    pub fn assign_r_type_arguments(
-        arguments: Vec<AstKind>,
-        args_to_use: &[ArgumentType],
-    ) -> AssembleResult<Self> {
-        let mut rd = 0;
-        let mut rs = 0;
-        let mut rt = 0;
-        let mut shamt = 0;
-
-        for (i, passed) in arguments.into_iter().enumerate() {
-            match args_to_use[i] {
-                ArgumentType::Rd | ArgumentType::Fd => {
-                    rd = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)?
-                }
-                ArgumentType::Rs | ArgumentType::Fs => {
-                    rs = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)?
-                }
-                ArgumentType::Rt | ArgumentType::Ft => {
-                    rt = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)?
-                }
-                ArgumentType::Immediate => {
-                    shamt = passed.get_immediate().ok_or(ErrorKind::InvalidArgument)?;
-                }
-                _ => unreachable!(),
-            }
-        }
-
-        // Bounds check on shift amount
-        if shamt > 31 {
-            return Err(ErrorKind::InvalidShamt);
-        }
-
-        return Ok(Self {
-            opcode: 0, // Will be filled in by the caller.
-            rd,
-            rs,
-            rt,
-            shamt,
-            funct: 0, // This will be filled in by the caller.
-        });
-    }
-}
-
-impl From<RawInstruction> for RArgs {
-    fn from(raw: RawInstruction) -> RArgs {
-        RArgs {
-            opcode: raw.get_opcode(),
-            rs: raw.get_rs(),
-            rt: raw.get_rt(),
-            rd: raw.get_rd(),
-            shamt: raw.get_shamt(),
-            funct: raw.get_funct(),
-        }
-    }
-}
 
 // FpCCArgs
-pub struct FpCCArgs {
-    pub opcode: u32,
-    pub fmt: u32,
-    pub ft: u32,
-    pub fs: u32,
-    pub cc: u32,
-    pub funct: u32,
-}
-
-impl FpCCArgs {
-    pub fn assign_fp_cc_arguments(
-        arguments: Vec<AstKind>,
-        args_to_use: &[ArgumentType],
-    ) -> AssembleResult<Self> {
-        let mut ft = 0;
-        let mut fs = 0;
-        let mut cc = 0;
-
-        for (i, passed) in arguments.into_iter().enumerate() {
-            match args_to_use[i] {
-                ArgumentType::Ft => {
-                    ft = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)? as u32;
-                }
-                ArgumentType::Fs => {
-                    fs = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)? as u32;
-                }
-                ArgumentType::Immediate => {
-                    if let AstKind::Immediate(num) = passed {
-                        // This may need more editing later should we end up implementing paired-single format.
-                        if num < 8 {
-                            cc = num;
-                        }
-                    }
-                }
-                _ => unreachable!(),
-            }
-        }
-
-        return Ok(Self {
-            opcode: 0, // Filled in by caller
-            fmt: 0,    // Filled in by caller
-            ft,
-            fs,
-            cc,
-            funct: 0, // Filled in by caller
-        });
-    }
-}
-
-impl From<RawInstruction> for FpCCArgs {
-    fn from(raw: RawInstruction) -> Self {
-        Self {
-            opcode: raw.get_opcode(),
-            fmt: raw.get_fmt(),
-            ft: raw.get_ft(),
-            fs: raw.get_fs(),
-            cc: raw.get_fd() >> 2,
-            funct: raw.get_funct(),
-        }
-    }
-}
 
 // FpCCBranchArgs
-/// The FpCCBranchArgs is for instructions like bc1t and bc1fl.
-/// The fields tf and nd are another layer of indirection to get the right instruction.
-#[derive(Debug)]
-pub struct FpCCBranchArgs {
-    pub opcode: u32,
-    pub funky_funct: u32, // The funct code is in a different place for this format. Little odd.
-    pub cc: u32,
-    pub tf: u32, // True/False
-    pub nd: u32, // Nullify delay slot (likely will set this bit to 1)
-    pub offset: u16,
-}
-
-impl FpCCBranchArgs {
-    pub fn assign_fp_cc_branch_arguments(
-        arguments: Vec<AstKind>,
-        args_to_use: &[ArgumentType],
-    ) -> AssembleResult<Self> {
-        let mut cc = 0;
-        let mut is_first_imm: bool = true; // This just keeps track of whether this is the first immediate we've encountered.
-        for (i, passed) in arguments.into_iter().enumerate() {
-            match args_to_use[i] {
-                ArgumentType::Immediate => {
-                    if is_first_imm {
-                        if let AstKind::Immediate(num) = passed {
-                            if num < 8 {
-                                cc = num;
-                            } else {
-                                return Err(ErrorKind::InvalidArgument);
-                            }
-                        }
-
-                        is_first_imm = false;
-                    } else {
-                        // Must be the second immediate, should go to zero.
-                        // Do nothing.
-                    }
-                }
-                ArgumentType::Identifier => (),
-                _ => todo!("Figure out what to do about bc1t cc, (imm) issue? maybe?"),
-            }
-        }
-
-        Ok(Self {
-            opcode: 0,
-            funky_funct: 0,
-            cc,
-            tf: 0,
-            nd: 0,
-            offset: 0,
-        })
-    }
-}
-
-impl From<RawInstruction> for FpCCBranchArgs {
-    fn from(raw: RawInstruction) -> Self {
-        Self {
-            opcode: raw.get_opcode(),
-            funky_funct: raw.get_fmt(),
-            cc: raw.get_ft() >> 2,
-            tf: raw.get_ft() & 1,
-            nd: (raw.get_ft() >> 1) & 1,
-            offset: raw.get_immediate(),
-        }
-    }
-}
 
 // FpRArgs
-pub struct FpRArgs {
-    pub opcode: u32,
-    pub fmt: u32,
-    pub ft: u32,
-    pub fs: u32,
-    pub fd: u32,
-    pub funct: u32,
-}
-
-impl FpRArgs {
-    pub fn assign_fp_r_arguments(
-        arguments: Vec<AstKind>,
-        args_to_use: &[ArgumentType],
-    ) -> AssembleResult<Self> {
-        let mut ft = 0;
-        let mut fs = 0;
-        let mut fd = 0;
-
-        for (i, passed) in arguments.into_iter().enumerate() {
-            match args_to_use[i] {
-                ArgumentType::Fd => {
-                    fd = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)? as u32
-                }
-                ArgumentType::Fs => {
-                    fs = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)? as u32
-                }
-                ArgumentType::Ft => {
-                    ft = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)? as u32
-                }
-                ArgumentType::Rt => {
-                    fd = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)? as u32
-                }
-                _ => unreachable!(),
-            }
-        }
-
-        return Ok(Self {
-            opcode: 0, // Will be filled in by caller
-            fmt: 0,    // Will be filled in by caller
-            ft,
-            fs,
-            fd,
-            funct: 0, // Will be filled in by caller
-        });
-    }
-}
-
-impl From<RawInstruction> for FpRArgs {
-    fn from(raw: RawInstruction) -> Self {
-        Self {
-            opcode: raw.get_opcode(),
-            fmt: raw.get_fmt(),
-            fs: raw.get_fs(),
-            ft: raw.get_ft(),
-            fd: raw.get_fd(),
-            funct: raw.get_funct(),
-        }
-    }
-}
 
 // FpFourRegArgs
-pub struct FpFourRegArgs {
-    pub op_code: u32,
-    pub fr: u32,
-    pub ft: u32,
-    pub fs: u32,
-    pub fd: u32,
-    pub op4: u32,
-    pub fmt3: u32,
-}
-
-impl FpFourRegArgs {
-    pub fn assign_fp_four_reg_arguments(
-        arguments: Vec<AstKind>,
-        args_to_use: &[ArgumentType],
-    ) -> AssembleResult<Self> {
-        let mut fr = 0;
-        let mut ft = 0;
-        let mut fs = 0;
-        let mut fd = 0;
-
-        for (i, passed) in arguments.into_iter().enumerate() {
-            match args_to_use[i] {
-                ArgumentType::Fr => fr = passed.get_register_as_u32().ok_or(ErrorKind::InvalidArgument)? as u32,
-                ArgumentType::Ft => ft = passed.get_register_as_u32().ok_or(ErrorKind::InvalidArgument)? as u32,
-                ArgumentType::Fs => fs = passed.get_register_as_u32().ok_or(ErrorKind::InvalidArgument)? as u32,
-                ArgumentType::Fd => fd = passed.get_register_as_u32().ok_or(ErrorKind::InvalidArgument)? as u32,
-                _ => unreachable!(),
-            }
-        }
-
-        return Ok(
-            Self {
-                op_code: 0, // Caller will fill this in
-                fr,
-                ft,
-                fs,
-                fd,
-                op4: 0, // Caller has this
-                fmt3: 0, // Caller has this
-            }
-        );
-    }
-}
-
-impl From<RawInstruction> for FpFourRegArgs {
-    fn from(raw: RawInstruction) -> Self {
-        Self {
-            op_code: (raw.raw >> 26) & 0b111111,
-            fr: (raw.raw >> 21) & 0b11111,
-            ft: (raw.raw >> 16) & 0b11111,
-            fs: (raw.raw >> 11) & 0b11111,
-            fd: (raw.raw >> 6) & 0b11111,
-            op4: (raw.raw >> 3) & 0b111,
-            fmt3: (raw.raw >> 3) & 0b111,
-        }
-    }
-}
 
 // RegImmIArgs
-#[derive(Debug)]
-pub struct RegImmIArgs {
-    pub op_code: u32,
-    pub rs: u32,
-    pub regimm_funct_code: u32,
-    pub imm: u32,
-}
-
-impl RegImmIArgs {
-    pub fn assign_regimm_i_arguments(
-        arguments: Vec<AstKind>,
-        args_to_use: &[ArgumentType],
-    ) -> AssembleResult<Self> {
-        let mut rs = 0;
-        let mut imm = 0;
-
-        for (i, passed) in arguments.into_iter().enumerate() {
-            match args_to_use[i] {
-                ArgumentType::Rs => {
-                    rs = passed
-                        .get_register_as_u32()
-                        .ok_or(ErrorKind::InvalidArgument)? as u32
-                }
-                ArgumentType::Immediate => {
-                    imm = passed.get_immediate().unwrap_or(0);
-                }
-                _ => unreachable!(),
-            }
-        }
-
-        return Ok(Self {
-            op_code: 0, // Will be filled in by caller
-            rs,
-            regimm_funct_code: 0, // Will be filled in by caller
-            imm,
-        });
-    }
-}
-
-impl From<RawInstruction> for RegImmIArgs {
-    fn from(raw: RawInstruction) -> Self {
-        RegImmIArgs {
-            op_code: raw.get_opcode(),
-            rs: raw.get_rs(),
-            regimm_funct_code: raw.get_rt(),
-            imm: raw.get_immediate() as u32,
-        }
-    }
-}
 
 // CopMovRArgs
-#[derive(Debug)]
-pub struct CopMovRArgs {
-    op_code: u32,
-    funct_code: u32,
-    rt: u32,
-    rd: u32,
-    sel: u32,
-}
-
-impl CopMovRArgs {
-    pub fn assign_cop_mov_arguments(
-        arguments: Vec<AstKind>,
-        args_to_use: &[ArgumentType]
-    ) -> AssembleResult<Self> {
-        let mut rt = 0;
-        let mut rd = 0;
-        let mut sel = 0;
-
-        for (i, passed) in arguments.into_iter().enumerate() {
-            match args_to_use[i] {
-                ArgumentType::Rt => rt = passed.get_register_as_u32().ok_or(ErrorKind::InvalidArgument)? as u32,
-                ArgumentType::Rd => rd = passed.get_register_as_u32().ok_or(ErrorKind::InvalidArgument)? as u32,
-                ArgumentType::Immediate => {
-                    let imm = passed.get_immediate().unwrap_or(0);
-                    if imm < 8 {
-                        sel = imm;
-                    } else {
-                        return Err(ErrorKind::InvalidArgument);
-                    }
-                },
-                _ => unreachable!(),
-            }
-        }
-
-        Ok(Self {
-            op_code: 0, // Will be filled in by caller
-            funct_code: 0,  // Will be filled in by caller
-            rt,
-            rd,
-            sel
-        })
-    }
-}
-
-impl From<RawInstruction> for CopMovRArgs {
-    fn from(raw: RawInstruction) -> Self {
-        Self {
-            op_code: (raw.raw >> 26) & 0b11_1111,
-            funct_code: (raw.raw >> 21) & 0b1_1111,
-            rt: (raw.raw >> 16) & 0b1_1111,
-            rd: (raw.raw >> 11) & 0b1_1111,
-            sel: raw.raw & 0b111
-        }
-    }
-}
